% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyseas-package.R
\docType{package}
\name{tidyseas-package}
\alias{tidyseas}
\alias{tidyseas-package}
\title{tidyseas: Tidy Tools for Pristine Seas Expedition Science}
\description{
Tools for the National Geographic Pristine Seas research team and its partners to carry expedition data from field sheets to the shared database to the report, the same way every time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pristine-seas/tidyseas}
  \item \url{https://pristine-seas.github.io/tidyseas/}
  \item Report bugs at \url{https://github.com/pristine-seas/tidyseas/issues}
}

}
\author{
\strong{Maintainer}: Juan Mayorga \email{marine.data.science@ngs.org}

}
\keyword{internal}
